feat: Test prepare_update_data
